requirements.txt
pyobjc==2.5.1
PyMouse==1.0
Pillow==3.2.0